+++ Cargo.toml
@@ -1,15 +1,17 @@
 [package]
 name = "dogstatsd"
-version = "0.6.2"
+version = "0.7.0"
 authors = ["Matt Casper <matthewvcasper@gmail.com>"]
 license = "MIT"
 description = "A DogstatsD client for Rust."
 documentation = "https://mcasper.github.io/dogstatsd-rs/dogstatsd/"
 repository = "https://github.com/mcasper/dogstatsd-rs"
 keywords = ["datadog", "dogstatsd", "client"]
+edition = "2018"
 
 [dependencies]
 chrono = "0.4"
+tokio = { version = "1.10.0", features = ["net", "rt", "macros"] }
 
 [features]
 unstable = []
