+++ Makefile
@@ -1,0 +1,5 @@
+compose:
+	docker-compose up -d
+
+compose-stop:
+	docker-compose stop
