+++ src/client/common.rs
@@ -1,0 +1,63 @@
+use crate::error::DogstatsdError;
+
+/// A type alias for returning a unit type or an error
+pub type DogstatsdResult = Result<(), DogstatsdError>;
+
+/// The struct that represents the options available for the Dogstatsd client.
+#[derive(Debug, PartialEq)]
+pub struct Options {
+    /// The address of the udp socket we'll bind to for sending.
+    pub from_addr: String,
+    /// The address of the udp socket we'll send metrics and events to.
+    pub to_addr: String,
+    /// A namespace to prefix all metrics with, joined with a '.'.
+    pub namespace: String,
+}
+
+impl Default for Options {
+    /// Create a new options struct with all the default settings.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    ///   use dogstatsd::Options;
+    ///
+    ///   let options = Options::default();
+    ///
+    ///   assert_eq!(
+    ///       Options {
+    ///           from_addr: "127.0.0.1:0".into(),
+    ///           to_addr: "127.0.0.1:8125".into(),
+    ///           namespace: String::new(),
+    ///       },
+    ///       options
+    ///   )
+    /// ```
+    fn default() -> Self {
+        Options {
+            from_addr: "127.0.0.1:0".into(),
+            to_addr: "127.0.0.1:8125".into(),
+            namespace: String::new(),
+        }
+    }
+
+}
+
+impl Options {
+    /// Create a new options struct by supplying values for all fields.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    ///   use dogstatsd::Options;
+    ///
+    ///   let options = Options::new("127.0.0.1:9000", "127.0.0.1:9001", "");
+    /// ```
+    pub fn new(from_addr: &str, to_addr: &str, namespace: &str) -> Self {
+        Options {
+            from_addr: from_addr.into(),
+            to_addr: to_addr.into(),
+            namespace: namespace.into(),
+        }
+    }
+}
